Assignment-3/requirements.txt
langchain
langchain-openai
pydantic
python-dotenv
openai